feature: refactor pin modal handling and separate modal logic into PinModal
